backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
twilio==9.0.0
phonenumbers==8.13.27
